Create index.js
The useEffect Hook allows us to trigger side effects when our components render. The effects can be run when specific data changes. We can define the dependencies of the effect to ensure React only runs it when the data it relies on changes.

There are two classifications of side effects. Some require cleanup, and others do not. We can return a function from our effect that describes how to clean up the side effect. Timers, event listeners and socket connections are examples of side effects that require cleanup.